+++ core-ui/src/components/Predefined/Details/ConfigMap/ConfigMap.details.js
@@ -27,6 +27,7 @@ export const ConfigMapsDetails = ({ DefaultRenderer, ...otherParams }) => {
     <DefaultRenderer
       customComponents={[ConfigMapEditor]}
       customColumns={customColumns}
+      resourceGraphProps={{ depth: 1 }}
       {...otherParams}
     />
   );
