+++ core-ui/src/components/Predefined/Details/CustomResourceDefinitions/CustomResourceDefinitions.details.js
@@ -112,6 +112,7 @@ export const CustomResourceDefinitionsDetails = ({
         },
       ]}
       {...otherParams}
+      resourceGraphProps={{ enabled: false }}
     />
   );
 };
