+++ core-ui/src/components/Predefined/Details/CustomResourceDefinitions/CustomResources.details.js
@@ -90,6 +90,7 @@ export function CustomResource({ params }) {
       breadcrumbs={breadcrumbs}
       customComponents={[AdditionalPrinterColumns]}
       i18n={i18n}
+      resourceGraphProps={{ enabled: false }}
     />
   );
 }
