+++ core-ui/src/components/Predefined/Details/Secret/Secret.details.js
@@ -41,6 +41,7 @@ export const SecretsDetails = ({ DefaultRenderer, ...otherParams }) => {
     <DefaultRenderer
       customComponents={[Secret, CertificateData, HelmReleaseDataWrapper]}
       customColumns={customColumns}
+      resourceGraphProps={{ depth: 1 }}
       {...otherParams}
     />
   );
