+++ core/src/i18n/en.yaml
@@ -443,6 +443,7 @@ common:
     delete: Delete
     edit: Edit
     edit-yaml: Edit YAML
+    load: Load
     ok: OK
     restart: Restart
     save: Save
@@ -548,7 +549,7 @@ common:
     duplicate-key: Duplicate key
     error: Error
     hide: Hide Read-Only Fields
-    k8s-name-input: "Name must consist of lowercase alphanumeric characters, can contain '-' (e.g.: 'my-name1')."
+    k8s-name-input: 'Name must consist of lowercase alphanumeric characters, can contain ''-'' (e.g.: ''my-name1'').'
     key: Key must start and end with an alphanumeric character, can contain '-', '_' or '.'.
     key-value: key=value, must start and end with an alphanumeric character, can contain '-', '_' or '.'.
     read-file: Read file content as a single value, with file name as a key.
@@ -1363,7 +1364,7 @@ jobs:
     complete: Job completed
     not-complete: Job in progress
 kubeconfig-id:
-  error: "Couldn't load kubeconfig ID; configuration not changed (Error: ${{error}})"
+  error: 'Couldn''t load kubeconfig ID; configuration not changed (Error: ${{error}})'
   must-be-an-object: Kubeconfig must be a JSON or YAML object.
 legal:
   copyright: Copyright
@@ -1664,6 +1665,11 @@ replica-sets:
   tooltips:
     error: At least one of the Pods has failed.
     running: All the Pods are running correctly.
+resource-graph:
+  relations:
+    exposed-by: Exposed by
+    exposes: Exposes
+  title: Resource Graph
 resource-quotas:
   headers:
     limits: Limits
