+++ shared/components/ResourceGraph/ResourceGraph.scss
@@ -1,0 +1,9 @@
+.react-flow {
+  height: 500px !important;
+  background: var(--sapNeutralBackground);
+}
+
+.resource-graph .fd-layout-panel__header {
+  display: flex;
+  justify-content: space-between;
+}
