+++ shared/components/ResourceGraph/getLayoutedElements.js
