+++ shared/components/ResourceGraph/helpers.js
