--- shared/components/graf.scss
@@ -1,4 +1,0 @@
-.react-flow {
-  height: 500px !important;
-  background: var(--sapNeutralBackground);
-}
